+++ e2e/specs/addressbook-tests.spec.js
@@ -10,8 +10,7 @@ import TabBarComponent from '../pages/TabBarComponent';
 import WalletActionsModal from '../pages/modals/WalletActionsModal';
 import AddAddressModal from '../pages/modals/AddAddressModal';
 
-import { loginToApp } from '../viewHelper';
-import { reverseServerPort } from '../helpers';
+import { loginToApp, reverseServerPort } from '../viewHelper';
 import FixtureBuilder from '../fixtures/fixture-builder';
 import {
   loadFixture,
@@ -26,7 +25,7 @@ const MEMO = 'Test adding ENS';
 
 describe(Smoke('Addressbook Tests'), () => {
   beforeAll(async () => {
-    await reverseServerPort(device);
+    await reverseServerPort();
     const fixture = new FixtureBuilder().build();
     await startFixtureServer();
     await loadFixture({ fixture });
