+++ e2e/specs/swaps/swap-tokens.spec.js
@@ -1,6 +1,6 @@
 'use strict';
 
-import { loginToApp } from '../../viewHelper';
+import { loginToApp, reverseServerPort } from '../../viewHelper';
 import { Regression } from '../../tags';
 import Onboarding from '../../pages/swaps/OnBoarding';
 import QuoteView from '../../pages/swaps/QuoteView';
@@ -18,12 +18,11 @@ import {
   stopFixtureServer,
 } from '../../fixtures/fixture-helper';
 import Networks from '../../resources/networks.json';
-import { reverseServerPort } from '../../utils';
 
 describe(Regression('Swap Tests'), () => {
   let swapOnboarded = false;
   beforeAll(async () => {
-    await reverseServerPort(device);
+    await reverseServerPort();
     const fixture = new FixtureBuilder()
       .withNetworkController(Networks.Tenderly)
       .build();
